==> 03-cleaning.Rmd <==
# Data transformation

For all three data sets `time_series_covid19_confirmed_global.csv`, `time_series_covid19_recovered_global.csv` and `time_series_covid19_deaths_global.csv`. As we are dealing with time series data, it makes sense for us to utilize the benefit of the tidy data format as we need both time and number of cases to be variables to plot. So we will first transformed the data sets into their tidy format. Specifically, we will collapse all columns with date names into different Date objects in a new variable named "Dates" and case values into a new variable that corresponds to date of the recorded cases. Variables Province/State, Country/Region, Lat, Long will not be changed. The tidy data are stored in `tidy_confirmed.csv`, `tidy_recovered.csv`, `tidy_deaths.csv` in the data folder. 

We also created a joined data sets from the tidy formats of all three data sets. We performed an outer join on variables Province/State, Country/Region, Lat, Long and Dates which are identical in all three data sets. This comprehensive tidy data set takes the union of all information from the three data sets which allows for more convenient comparisons of confirmed cases, recovered cases and deaths given query countries. It also allows faster access to each country's case record without having to access all three data sets. The full joined data set is stored in `joined_df.csv` in the data folder. 

In addition, since our data records daily cumulative cases for each country from 2020-01-22 to 2020-11-21, we can obtain the number of cumulative cases at the end of month in the month from January to October in 2020 for all 269 countries/regions. This observation allows us create many visualization that explores the distribution or the growth of cases either in individual country or globally. For example, we can obtain the global monthly total confirmed cases which is a high level overview of the spread of Covid-19 in the globe. (we actually use this in our interactive part). 

```{r}
library(tidyverse)
library(plyr)

#read in data sets 
confirmed = read.csv("data/time_series_covid19_confirmed_global.csv", check.names = FALSE)
recovered = read.csv("data/time_series_covid19_recovered_global.csv", check.names = FALSE)
deaths = read.csv("data/time_series_covid19_deaths_global.csv", check.names = FALSE)

format_dates = function(x){
  to_date = as.Date(x, "%m/%d/%y")
  return(to_date)
}

#pivot datasets into tidy data and format dates as Data objects
tidy_confirmed = confirmed %>% 
  pivot_longer(cols = !c("Province/State", "Country/Region", "Lat", "Long"), 
               names_to = "Dates", values_to = "Confirmed", names_transform = list(Date = format_dates)) 
write.csv(tidy_confirmed,"data/tidy_confirmed.csv", row.names = FALSE)

tidy_recovered = recovered %>% 
  pivot_longer(cols = !c("Province/State", "Country/Region", "Lat", "Long"), 
               names_to = "Dates", values_to = "Recovered", names_transform = list(Date = format_dates)) 
write.csv(tidy_recovered,"data/tidy_recovered.csv", row.names = FALSE)

tidy_deaths = deaths %>% 
  pivot_longer(cols = !c("Province/State", "Country/Region", "Lat", "Long"), 
               names_to = "Dates", values_to = "Deaths", names_transform = list(Date = format_dates))
write.csv(tidy_deaths,"data/tidy_deaths.csv", row.names = FALSE)

#join data sets
df = join(tidy_confirmed, tidy_recovered)
df = join(df, tidy_deaths)
write.csv(df,"data/joined_df.csv", row.names = FALSE)
```

```{r}
#retrieve end of month dates
month_end = c("1/31/20","2/29/20", "3/31/20", "4/30/20", "5/31/20", "6/30/20", "7/31/20", "8/31/20", "9/30/20", "10/31/20")


monthly_df = confirmed[c("Province/State", "Country/Region", month_end)]
tidy_month = monthly_df %>% pivot_longer(!c(1,2), names_to = "Dates", names_transform = list(Date = format_dates), values_to = "Cases")

month_end_date = seq(as.Date("2020-02-01"),length=10,by="months")-1
month_total = apply(monthly_df[month_end], 2, sum)
```
