==> 06-interactive.Rmd <==
# Interactive component

<script src="https://d3js.org/d3.v6.min.js"></script>

<button id="start" style="position: absolute; margin-left: 6px; margin-top: 4px;">Start Animation</button>
<button id="reset" style="position: absolute; margin-left: 150px; margin-top: 4px">Reset</button>

<div id="plot"></div>

<script src="scripts/draw_total_cases.js"></script>

In this project, we created a nice animation that plots the number of monthly confirmed cases around the globe in 2020 as it increases monthly! It draws attention to the trend of growth of the confirmed cases as the red line plots the number of cases in real time! Specifically, with months from January to October on the axis and total confirmed cases globally (in thousands), the line is undoubtedly going up continuously since January this year. We also set up a case counter that reflects to the monthly cases in each individual cases as the time series line passes through the months. You can use the "Start Animation" to start the animation while "Reset" bottoms removes the line and case counter to reset. 

The motivation for this simple yet meaningful animation is that, compared to static time series line chart of daily cases at the country level, following along with the growth or decrease of monthly global confirmed cases enables us to be more sensitive to fluctuation at a larger scale. It allows to detect any significant pattern changes in general, ignoring minor daily fluctuations. This could be useful for historical purposes since it captures the general trend of total global cases over the months.

Our D3 visualization using tooltip is inspired by Mike Foster and Eric Huntley's work http://duspviz.mit.edu/d3-workshop/transitions-animation/. 

Side Note, there is a little issue with the x scale on the x axis as it appears to have shifted to the left. It is a continuing issue for our interactive visualization in this chapter. Feel free to help us out!
