==> 07-conclusion.Rmd <==
# Conclusion

In this project, we use time series line charts, map, grouped bar chart and Cleveland dot plot to visualize the impact of Covid-19 pandemic. Specifically, we select representative countries/regions that top the charts to explore. This is due to the limitation of certain graphs that cannot accommodate two many variables. For example, we selected the top 5 countries to plot in time series line charts in terms of their confirmed cases, recovered cases and deaths. Choosing a few number of selective countries/regions, instead of all 269 countries, to plot will avoid clustering/overlapping in the graph which can be very messy. However, this is a trade-off between generality and information about other countries/regions. 

One possible improvement is to group countries/regions by their longitude and latitude. Specifically, we can build grid on the world map to group countries that are close to each other geographically. However, while it allows for less cluttering, it might be over-generalize and give misleading geographical patterns.    

Another limitation is that the depth of our visualization is confined by the data source as we only have information about the number of cases and simple geographical information about each country. If we have more information such as population density of each countries, then we might be able to visualize proportion of Covid-19 cases to population for each country which is a better indicator of the pandemic spread on the country/region level.
